Move delay to scheduler & ctx
