+++ zetaclient/chains/evm/observer/observer_test.go
@@ -71,6 +71,7 @@ func getAppContext(
 		nil,
 		chainParams,
 		*sample.CrosschainFlags(),
+		sample.OperationalFlags(),
 	)
 	require.NoError(t, err)
