+++ zetaclient/chains/evm/signer/signer_test.go
@@ -211,6 +211,7 @@ func makeCtx(t *testing.T) context.Context {
 			chains.BscMainnet.ChainId: &bscParams,
 		},
 		observertypes.CrosschainFlags{},
+		observertypes.OperationalFlags{},
 	)
 	require.NoError(t, err, "unable to update app context")
